requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
geoalchemy2==0.14.2
python-dotenv==1.0.0
qdrant-client==1.6.4
sentence-transformers==2.2.2
pydantic==2.5.2
pytest==7.4.3
alembic==1.12.1
python-multipart==0.0.6
shapely==2.0.2 